requirements.txt
joblib==1.1.0
numpy==1.19.5
pandas==1.2.5
Pillow==9.2.0
streamlit==1.12.2
scikit-learn==1.0.2
jupyter==1.0.0